requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic>=0.40.0
qdrant-client==1.7.3
pyodbc==5.0.1
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.2.0